+++ nautilus_trader/data/client.pyx
@@ -519,7 +519,7 @@ cdef class MarketDataClient(DataClient):
 
     cpdef void subscribe_venue_status_updates(self, Venue venue):
         """
-        Subscribe to `InstrumentStatusUpdate` data for the venue.
+        Subscribe to `InstrumentStatus` data for the venue.
 
         Parameters
         ----------
@@ -528,14 +528,14 @@ cdef class MarketDataClient(DataClient):
 
         """
         self._log.error(  # pragma: no cover
-            f"Cannot subscribe to `VenueStatusUpdate` data for {venue}: not implemented. "  # pragma: no cover
+            f"Cannot subscribe to `VenueStatus` data for {venue}: not implemented. "  # pragma: no cover
             f"You can implement by overriding the `subscribe_venue_status_updates` method for this client.",  # pragma: no cover
         )
         raise NotImplementedError("method must be implemented in the subclass")
 
     cpdef void subscribe_instrument_status_updates(self, InstrumentId instrument_id):
         """
-        Subscribe to `InstrumentStatusUpdates` data for the given instrument ID.
+        Subscribe to `InstrumentStatus` data for the given instrument ID.
 
         Parameters
         ----------
@@ -544,7 +544,7 @@ cdef class MarketDataClient(DataClient):
 
         """
         self._log.error(  # pragma: no cover
-            f"Cannot subscribe to `InstrumentStatusUpdates` data for {instrument_id}: not implemented. "  # pragma: no cover
+            f"Cannot subscribe to `InstrumentStatus` data for {instrument_id}: not implemented. "  # pragma: no cover
             f"You can implement by overriding the `subscribe_instrument_status_updates` method for this client.",  # pragma: no cover
         )
         raise NotImplementedError("method must be implemented in the subclass")
@@ -721,7 +721,7 @@ cdef class MarketDataClient(DataClient):
 
     cpdef void unsubscribe_venue_status_updates(self, Venue venue):
         """
-        Unsubscribe from `InstrumentStatusUpdate` data for the given venue.
+        Unsubscribe from `InstrumentStatus` data for the given venue.
 
         Parameters
         ----------
@@ -730,14 +730,14 @@ cdef class MarketDataClient(DataClient):
 
         """
         self._log.error(  # pragma: no cover
-            f"Cannot unsubscribe from `VenueStatusUpdates` data for {venue}: not implemented. "  # pragma: no cover
+            f"Cannot unsubscribe from `VenueStatus` data for {venue}: not implemented. "  # pragma: no cover
             f"You can implement by overriding the `unsubscribe_venue_status_updates` method for this client.",  # pragma: no cover
         )
         raise NotImplementedError("method must be implemented in the subclass")
 
     cpdef void unsubscribe_instrument_status_updates(self, InstrumentId instrument_id):
         """
-        Unsubscribe from `InstrumentStatusUpdate` data for the given instrument ID.
+        Unsubscribe from `InstrumentStatus` data for the given instrument ID.
 
         Parameters
         ----------
@@ -746,7 +746,7 @@ cdef class MarketDataClient(DataClient):
 
         """
         self._log.error(  # pragma: no cover
-            f"Cannot unsubscribe from `InstrumentStatusUpdates` data for {instrument_id}: not implemented. "  # pragma: no cover
+            f"Cannot unsubscribe from `InstrumentStatus` data for {instrument_id}: not implemented. "  # pragma: no cover
             f"You can implement by overriding the `unsubscribe_instrument_status_updates` method for this client.",  # pragma: no cover
         )
         raise NotImplementedError("method must be implemented in the subclass")