+++ nautilus_trader/model/data/status.pyx
@@ -28,17 +28,7 @@ from nautilus_trader.model.identifiers cimport Venue
 from nautilus_trader.model.objects cimport Price
 
 
-cdef class StatusUpdate(Data):
-    """
-    The base class for all status updates.
-
-    Warnings
-    --------
-    This class should not be used directly, but through a concrete subclass.
-    """
-
-
-cdef class VenueStatusUpdate(StatusUpdate):
+cdef class VenueStatus(Data):
     """
     Represents an update that indicates a change in a Venue status.
 
@@ -66,11 +56,11 @@ cdef class VenueStatusUpdate(StatusUpdate):
         self.ts_event = ts_event
         self.ts_init = ts_init
 
-    def __eq__(self, VenueStatusUpdate other) -> bool:
-        return VenueStatusUpdate.to_dict_c(self) == VenueStatusUpdate.to_dict_c(other)
+    def __eq__(self, VenueStatus other) -> bool:
+        return VenueStatus.to_dict_c(self) == VenueStatus.to_dict_c(other)
 
     def __hash__(self) -> int:
-        return hash(frozenset(VenueStatusUpdate.to_dict_c(self)))
+        return hash(frozenset(VenueStatus.to_dict_c(self)))
 
     def __repr__(self) -> str:
         return (
@@ -80,9 +70,9 @@ cdef class VenueStatusUpdate(StatusUpdate):
         )
 
     @staticmethod
-    cdef VenueStatusUpdate from_dict_c(dict values):
+    cdef VenueStatus from_dict_c(dict values):
         Condition.not_none(values, "values")
-        return VenueStatusUpdate(
+        return VenueStatus(
             venue=Venue(values["venue"]),
             status=market_status_from_str(values["status"]),
             ts_event=values["ts_event"],
@@ -90,10 +80,10 @@ cdef class VenueStatusUpdate(StatusUpdate):
         )
 
     @staticmethod
-    cdef dict to_dict_c(VenueStatusUpdate obj):
+    cdef dict to_dict_c(VenueStatus obj):
         Condition.not_none(obj, "obj")
         return {
-            "type": "VenueStatusUpdate",
+            "type": "VenueStatus",
             "venue": obj.venue.to_str(),
             "status": market_status_to_str(obj.status),
             "ts_event": obj.ts_event,
@@ -101,7 +91,7 @@ cdef class VenueStatusUpdate(StatusUpdate):
         }
 
     @staticmethod
-    def from_dict(dict values) -> VenueStatusUpdate:
+    def from_dict(dict values) -> VenueStatus:
         """
         Return a venue status update from the given dict values.
 
@@ -112,13 +102,13 @@ cdef class VenueStatusUpdate(StatusUpdate):
 
         Returns
         -------
-        VenueStatusUpdate
+        VenueStatus
 
         """
-        return VenueStatusUpdate.from_dict_c(values)
+        return VenueStatus.from_dict_c(values)
 
     @staticmethod
-    def to_dict(VenueStatusUpdate obj):
+    def to_dict(VenueStatus obj):
         """
         Return a dictionary representation of this object.
 
@@ -127,10 +117,10 @@ cdef class VenueStatusUpdate(StatusUpdate):
         dict[str, object]
 
         """
-        return VenueStatusUpdate.to_dict_c(obj)
+        return VenueStatus.to_dict_c(obj)
 
 
-cdef class InstrumentStatusUpdate(StatusUpdate):
+cdef class InstrumentStatus(Data):
     """
     Represents an event that indicates a change in an instrument status.
 
@@ -158,11 +148,11 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
         self.ts_event = ts_event
         self.ts_init = ts_init
 
-    def __eq__(self, InstrumentStatusUpdate other) -> bool:
-        return InstrumentStatusUpdate.to_dict_c(self) == InstrumentStatusUpdate.to_dict_c(other)
+    def __eq__(self, InstrumentStatus other) -> bool:
+        return InstrumentStatus.to_dict_c(self) == InstrumentStatus.to_dict_c(other)
 
     def __hash__(self) -> int:
-        return hash(frozenset(InstrumentStatusUpdate.to_dict_c(self)))
+        return hash(frozenset(InstrumentStatus.to_dict_c(self)))
 
     def __repr__(self) -> str:
         return (
@@ -172,9 +162,9 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
         )
 
     @staticmethod
-    cdef InstrumentStatusUpdate from_dict_c(dict values):
+    cdef InstrumentStatus from_dict_c(dict values):
         Condition.not_none(values, "values")
-        return InstrumentStatusUpdate(
+        return InstrumentStatus(
             instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
             status=market_status_from_str(values["status"]),
             ts_event=values["ts_event"],
@@ -182,10 +172,10 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
         )
 
     @staticmethod
-    cdef dict to_dict_c(InstrumentStatusUpdate obj):
+    cdef dict to_dict_c(InstrumentStatus obj):
         Condition.not_none(obj, "obj")
         return {
-            "type": "InstrumentStatusUpdate",
+            "type": "InstrumentStatus",
             "instrument_id": obj.instrument_id.to_str(),
             "status": market_status_to_str(obj.status),
             "ts_event": obj.ts_event,
@@ -193,7 +183,7 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
         }
 
     @staticmethod
-    def from_dict(dict values) -> InstrumentStatusUpdate:
+    def from_dict(dict values) -> InstrumentStatus:
         """
         Return an instrument status update from the given dict values.
 
@@ -204,13 +194,13 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
 
         Returns
         -------
-        InstrumentStatusUpdate
+        InstrumentStatus
 
         """
-        return InstrumentStatusUpdate.from_dict_c(values)
+        return InstrumentStatus.from_dict_c(values)
 
     @staticmethod
-    def to_dict(InstrumentStatusUpdate obj):
+    def to_dict(InstrumentStatus obj):
         """
         Return a dictionary representation of this object.
 
@@ -219,7 +209,7 @@ cdef class InstrumentStatusUpdate(StatusUpdate):
         dict[str, object]
 
         """
-        return InstrumentStatusUpdate.to_dict_c(obj)
+        return InstrumentStatus.to_dict_c(obj)
 
 
 cdef class InstrumentClose(Data):