+++ nautilus_trader/test_kit/stubs/data.py
@@ -31,13 +31,13 @@ from nautilus_trader.model.data import BarSpecification
 from nautilus_trader.model.data import BarType
 from nautilus_trader.model.data import BookOrder
 from nautilus_trader.model.data import InstrumentClose
-from nautilus_trader.model.data import InstrumentStatusUpdate
+from nautilus_trader.model.data import InstrumentStatus
 from nautilus_trader.model.data import OrderBookDelta
 from nautilus_trader.model.data import OrderBookDeltas
 from nautilus_trader.model.data import QuoteTick
 from nautilus_trader.model.data import Ticker
 from nautilus_trader.model.data import TradeTick
-from nautilus_trader.model.data import VenueStatusUpdate
+from nautilus_trader.model.data import VenueStatus
 from nautilus_trader.model.enums import AggressorSide
 from nautilus_trader.model.enums import BarAggregation
 from nautilus_trader.model.enums import BookAction
@@ -390,8 +390,8 @@ class TestDataStubs:
     def venue_status_update(
         venue: Venue | None = None,
         status: MarketStatus | None = None,
-    ) -> VenueStatusUpdate:
-        return VenueStatusUpdate(
+    ) -> VenueStatus:
+        return VenueStatus(
             venue=venue or Venue("BINANCE"),
             status=status or MarketStatus.OPEN,
             ts_event=0,
@@ -402,8 +402,8 @@ class TestDataStubs:
     def instrument_status_update(
         instrument_id: InstrumentId | None = None,
         status: MarketStatus | None = None,
-    ) -> InstrumentStatusUpdate:
-        return InstrumentStatusUpdate(
+    ) -> InstrumentStatus:
+        return InstrumentStatus(
             instrument_id=instrument_id or InstrumentId(Symbol("BTCUSDT"), Venue("BINANCE")),
             status=status or MarketStatus.PAUSE,
             ts_event=0,