+++ RELEASES.md
@@ -19,6 +19,9 @@ This will be the final release with support for Python 3.9.
 
 ### Breaking Changes
 - Renamed `BookType.L1_TBBO` to `BookType.L1_MBP` (more accurate definition, as L1 is the top-level price either side)
+- Renamed `VenueStatusUpdate` -> `VenueStatus`
+- Renamed `InstrumentStatusUpdate` -> `InstrumentStatus`
+- Changed `InstrumentStatus` fields/schema and constructor
 - Moved `manage_gtd_expiry` from `Strategy.submit_order(...)` and `Strategy.submit_order_list(...)` to `StrategyConfig` (simpler and allows re-activiting any GTD timers on start)
 
 ### Fixes